+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Ability to build Kokkos and RAJA versions against existing packages.
 - Thrust managed memory.
+- HIP managed memory.
+- New implementation using SYCL2020 USM (sycl2020-acc) and renamed original `sycl2020` to `sycl2020-acc`.
 
 ### Changed
 - RAJA CUDA CMake build issues resolved.
@@ -13,6 +15,10 @@ All notable changes to this project will be documented in this file.
 - OneAPI DPCPP compiler is deprecated in favour of ICPX, so added new build option to SYCL 2020 version.
 - Updates to the HIP kernels and API usage.
 - Number of thread-blocks in CUDA dot kernel implementation changed to 1024.
+- Fix compatibility of `sycl2020` (now `sycl2020-acc`) with hipSYCL. 
+- Bumped Julia compat to 1.9
+- Bumped Rust to 1.74.0-nightly (13e6f24b9 2023-09-23)
+   
 
 ## [v4.0] - 2021-12-22